refactor: 프로젝트 화면 헤더 컴포넌트 분리
