+++ src/components/projects/header/ManageFileSidebar.vue
