+++ src/components/projects/header/SearchRequirementsSidebar.vue
